First review pages up

+++ createdb.rb
@@ -14,7 +14,7 @@ DB.create_table! :reviews do
   primary_key :id
   foreign_key :book_id
   foreign_key :user_id
-  Numeric :rating
+  Integer :rating
   String :comments, text: true
 end
 DB.create_table! :users do
